feat: OutputView Car 관련 정보 삭제 및 결과 정보를 만드는 CarStatus 추가
